dw/requirements.txt
#Dependencies used in this library. Must be ran in linux since it uses cPython.
#pip install -r requirements.txt
mysql-connector-python==8.3.0